+++ cmd/timetracker-tray/main.go
@@ -3,22 +3,24 @@ package main
 import (
 	"flag"
 	"fmt"
+	"os"
+	"runtime"
 
 	"github.com/neflyte/timetracker/cmd/timetracker-tray/cmd"
 	"github.com/neflyte/timetracker/lib/constants"
+	"github.com/neflyte/timetracker/lib/logger"
 	"github.com/neflyte/timetracker/lib/startup"
 	"github.com/neflyte/timetracker/lib/ui/tray"
+	"github.com/neflyte/timetracker/lib/utils"
 )
 
-/*const (
-	trayPidfile = "timetracker-tray.pid"
-)*/
-
 var (
-	configFileName string
-	logLevel       string
-	showVersion    bool
-	console        bool
+	configFileName       string
+	logLevel             string
+	showVersion          bool
+	console              bool
+	darwinEnableLaunchd  bool
+	darwinDisableLaunchd bool
 )
 
 func init() {
@@ -26,6 +28,11 @@ func init() {
 	flag.StringVar(&logLevel, "logLevel", constants.DefaultLogLevel, "Specify the logging level")
 	flag.BoolVar(&showVersion, "version", false, "Display the program version")
 	flag.BoolVar(&console, "console", false, "Log to the console")
+	// macOS-only flags to manage launchd
+	if runtime.GOOS == "darwin" {
+		flag.BoolVar(&darwinEnableLaunchd, "enableLaunchd", false, "Start Timetracker Tray at login")
+		flag.BoolVar(&darwinDisableLaunchd, "disableLaunchd", false, "Do not start Timetracker Tray at login")
+	}
 }
 
 func main() {
@@ -34,26 +41,44 @@ func main() {
 		fmt.Printf("timetracker-tray %s\n", cmd.AppVersion)
 		return
 	}
+
 	startup.SetLogLevel(logLevel)
 	startup.SetConsole(console)
 	startup.InitLogger()
 	defer startup.CleanupLogger()
+
+	if runtime.GOOS == "darwin" {
+		doDarwinLaunchdCommands(darwinEnableLaunchd, darwinDisableLaunchd)
+	}
+
 	startup.SetDatabaseFileName(configFileName)
 	startup.InitDatabase()
 	defer startup.CleanupDatabase()
-	/*log := logger.GetLogger("main")
-	err := preDoTray()
-	if err != nil {
-		log.Err(err).
-			Msg("error setting up tray entry")
-		return
-	}
-	defer func() {
-		err = postDoTray()
-		if err != nil {
-			log.Err(err).
-				Msg("error tearing down tray entry")
-		}
-	}()*/
 	tray.Run(nil)
+}
+
+func doDarwinLaunchdCommands(enable bool, disable bool) {
+	log := logger.GetLogger("doDarwinLaunchdCommands")
+	if enable && disable {
+		log.Fatal().
+			Msg("Cannot specify both -enableLaunchd and -disableLaunchd at the same time")
+	}
+	if enable {
+		err := utils.EnableLaunchd()
+		if err != nil {
+			log.Fatal().
+				Err(err).
+				Msg("Cannot start Timetracker Tray at login")
+		}
+		os.Exit(0)
+	}
+	if disable {
+		err := utils.DisableLaunchd()
+		if err != nil {
+			log.Fatal().
+				Err(err).
+				Msg("Cannot disable Timetracker Tray at login")
+		}
+		os.Exit(0)
+	}
 }